前後端優化

+++ backend/app/models/stock_daily_price.py
@@ -1,5 +1,5 @@
 from datetime import date
-from sqlalchemy import String, Date, Numeric
+from sqlalchemy import String, Date, Float
 from sqlalchemy.orm import Mapped, mapped_column
 from app.db.database import Base
 
@@ -9,6 +9,8 @@ class StockDailyPrice(Base):
     id: Mapped[int] = mapped_column(primary_key=True, autoincrement=True)
     symbol: Mapped[str] = mapped_column(String, index=True)
     trade_date: Mapped[date] = mapped_column(Date, index=True)
-    close_price: Mapped[float] = mapped_column(Numeric)
-    change_percent: Mapped[float] = mapped_column(Numeric)
-    volume: Mapped[float | None] = mapped_column(Numeric, nullable=True)
+    # Float (not Numeric): PostgreSQL returns Decimal for Numeric, which breaks
+    # float arithmetic in the calculators; these are display values, not ledger money
+    close_price: Mapped[float] = mapped_column(Float)
+    change_percent: Mapped[float] = mapped_column(Float)
+    volume: Mapped[float | None] = mapped_column(Float, nullable=True)
